chore: improve loading experience and up page size
