+++ packages/frontend/src/features/metricsCatalog/hooks/useCatalogCategories.tsx
@@ -2,7 +2,6 @@ import {
     type ApiError,
     type ApiMetricsCatalog,
     type ApiSuccessEmpty,
-    type CatalogItemIcon,
     type Tag,
 } from '@lightdash/common';
 import {
@@ -44,6 +43,7 @@ export const useAddCategoryToCatalogItem = () => {
             previousCatalog: unknown;
         }
     >({
+        mutationKey: ['add-category'],
         mutationFn: addCategoryToCatalogItem,
         onMutate: async ({ catalogSearchUuid, tagUuid, projectUuid }) => {
             // Cancel any outgoing refetches
@@ -138,6 +138,7 @@ export const useRemoveCategoryFromCatalogItem = () => {
             previousCatalog: unknown;
         }
     >({
+        mutationKey: ['remove-category'],
         mutationFn: removeCategoryFromCatalogItem,
         onMutate: async ({ catalogSearchUuid, tagUuid, projectUuid }) => {
             // Cancel any outgoing refetches
@@ -165,88 +166,6 @@ export const useRemoveCategoryFromCatalogItem = () => {
                             item.categories = item.categories.filter(
                                 (category) => category.tagUuid !== tagUuid,
                             );
-                        },
-                        catalogSearchUuid,
-                    ),
-            );
-
-            return { previousCatalog };
-        },
-        onError: (_, __, context) => {
-            if (context?.previousCatalog) {
-                Object.entries(context.previousCatalog).forEach(
-                    ([queryKeyStr, data]) => {
-                        const queryKey = JSON.parse(queryKeyStr);
-                        queryClient.setQueryData<
-                            InfiniteData<ApiMetricsCatalog['results']>
-                        >(queryKey, data);
-                    },
-                );
-            }
-        },
-        onSettled: (_, __, { projectUuid }) => {
-            void queryClient.invalidateQueries([
-                'metrics-catalog',
-                projectUuid,
-            ]);
-        },
-    });
-};
-
-type UpdateCatalogItemIconParams = {
-    projectUuid: string;
-    catalogSearchUuid: string;
-    icon: CatalogItemIcon | null;
-};
-
-const updateCatalogItemIcon = async ({
-    projectUuid,
-    catalogSearchUuid,
-    icon,
-}: UpdateCatalogItemIconParams) => {
-    return lightdashApi<ApiSuccessEmpty['results']>({
-        url: `/projects/${projectUuid}/dataCatalog/${catalogSearchUuid}/icon`,
-        method: 'PATCH',
-        body: JSON.stringify({ icon }),
-    });
-};
-
-/**
- * Update a catalog item's icon
- */
-export const useUpdateCatalogItemIcon = () => {
-    const queryClient = useQueryClient();
-    return useMutation<
-        ApiSuccessEmpty['results'],
-        ApiError,
-        UpdateCatalogItemIconParams,
-        {
-            previousCatalog: unknown;
-        }
-    >({
-        mutationFn: updateCatalogItemIcon,
-        onMutate: async ({ catalogSearchUuid, icon, projectUuid }) => {
-            await queryClient.cancelQueries({
-                queryKey: ['metrics-catalog', projectUuid],
-            });
-
-            const previousCatalog = queryClient.getQueryData([
-                'metrics-catalog',
-                projectUuid,
-            ]);
-
-            queryClient.setQueriesData<
-                InfiniteData<ApiMetricsCatalog['results']>
-            >(
-                {
-                    queryKey: ['metrics-catalog', projectUuid],
-                    exact: false,
-                },
-                (old) =>
-                    updateMetricsCatalogQuery(
-                        old,
-                        (item) => {
-                            item.icon = icon;
                         },
                         catalogSearchUuid,
                     ),
