+++ packages/frontend/src/features/metricsCatalog/hooks/useProjectTags.tsx
@@ -45,6 +45,7 @@ export const useCreateTag = () => {
             previousCatalog: unknown;
         }
     >({
+        mutationKey: ['create-tag'],
         mutationFn: async ({ projectUuid, data, catalogSearchUuid }) => {
             const newTag = await createTag(projectUuid, data);
             if (catalogSearchUuid) {
@@ -189,6 +190,7 @@ export const useUpdateTag = () => {
             previousCatalog: unknown;
         }
     >({
+        mutationKey: ['update-tag'],
         mutationFn: ({ projectUuid, tagUuid, data }) => {
             return updateTag(projectUuid, tagUuid, data);
         },
@@ -277,6 +279,7 @@ export const useDeleteTag = () => {
             previousCatalog: unknown;
         }
     >({
+        mutationKey: ['delete-tag'],
         mutationFn: ({ projectUuid, tagUuid }) => {
             return deleteTag(projectUuid, tagUuid);
         },
